refactor: 노선 생성자 추가
